Factored out String interner into a different module.
